% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_q1.R
\name{boxplot_q1}
\alias{boxplot_q1}
\title{question1: boxplot}
\usage{
boxplot_q1(data_1, ydata)
}
\arguments{
\item{data_1}{The data we want to input, here we use the DRG data}

\item{ydata}{Here we choose from " Average Covered Charges "," Average Total Payments","Average Medicare Payments" as the data of y axis}
}
\value{
a box plot with x axis as the DRG code(only keep the number part) and y axis as the ydata selected from three choices
}
\description{
question1: boxplot
}
\examples{
boxplot_q1(df,ydata = 'Average Covered Charges')
}
