% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_q2.R
\name{stat_q2}
\alias{stat_q2}
\title{question2:calculation of mean,median and standard deviation}
\usage{
stat_q2(df, x)
}
\arguments{
\item{df}{The data we want to input, here we use the DRG data}

\item{x}{The condition argument, we input "Mean","Standard deviation","Median" to get back corresponding value.}
}
\value{
a dataframe with the corresponding data we choose by x
}
\description{
question2:calculation of mean,median and standard deviation
}
\examples{
lab2_2(df,x='Mean')
}
